# build: replace webpack with rspack