--- a/Makefile
+++ b/Makefile
@@ -35,7 +35,6 @@ setup-pyenv:
 
 build-js-po: node-version-check
 	mkdir -p build
-	rm -rf node_modules/.cache/babel-loader
 	SENTRY_EXTRACT_TRANSLATIONS=1 $(WEBPACK)
 
 build-spectacular-docs: