--- a/rspack.config.ts
+++ b/rspack.config.ts
@@ -4,36 +4,22 @@
 import fs from 'fs';
 import path from 'path';
 
-import {WebpackReactSourcemapsPlugin} from '@acemarke/react-prod-sourcemaps';
+import rspack from '@rspack/core';
+import ReactRefreshRspackPlugin from '@rspack/plugin-react-refresh';
 import CompressionPlugin from 'compression-webpack-plugin';
-import CopyPlugin from 'copy-webpack-plugin';
-import CssMinimizerPlugin from 'css-minimizer-webpack-plugin';
-import ForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
-import MiniCssExtractPlugin from 'mini-css-extract-plugin';
-import webpack from 'webpack';
-import {Configuration as DevServerConfig} from 'webpack-dev-server';
+import HtmlWebpackPlugin from 'html-webpack-plugin';
+// TODO(@anonrig): Remove this when Rspack bundles it.
+import {ContextReplacementPlugin} from 'webpack';
 import WebpackHookPlugin from 'webpack-hook-plugin';
-import FixStyleOnlyEntriesPlugin from 'webpack-remove-empty-scripts';
 
-import IntegrationDocsFetchPlugin from './build-utils/integration-docs-fetch-plugin';
 import LastBuiltPlugin from './build-utils/last-built-plugin';
 import SentryInstrumentation from './build-utils/sentry-instrumentation';
 import {extractIOSDeviceNames} from './scripts/extract-ios-device-names';
-import babelConfig from './babel.config';
 
 // Runs as part of prebuild step to generate a list of identifier -> name mappings for  iOS
 (async () => {
   await extractIOSDeviceNames();
 })();
-
-/**
- * Merges the devServer config into the webpack config
- *
- * See: https://github.com/DefinitelyTyped/DefinitelyTyped/issues/43232
- */
-interface Configuration extends webpack.Configuration {
-  devServer?: DevServerConfig;
-}
 
 const {env} = process;
 
@@ -57,7 +43,9 @@ const IS_ACCEPTANCE_TEST = !!env.IS_ACCEPTANCE_TEST;
 const IS_DEPLOY_PREVIEW = !!env.NOW_GITHUB_DEPLOYMENT;
 const IS_UI_DEV_ONLY = !!env.SENTRY_UI_DEV_ONLY;
 const DEV_MODE = !(IS_PRODUCTION || IS_CI);
-const WEBPACK_MODE: Configuration['mode'] = IS_PRODUCTION ? 'production' : 'development';
+const WEBPACK_MODE: rspack.Configuration['mode'] = IS_PRODUCTION
+  ? 'production'
+  : 'development';
 const CONTROL_SILO_PORT = env.SENTRY_CONTROL_SILO_PORT;
 
 // Environment variables that are used by other tooling and should
@@ -76,9 +64,9 @@ const HAS_WEBPACK_DEV_SERVER_CONFIG =
 
 // User/tooling configurable environment variables
 const NO_DEV_SERVER = !!env.NO_DEV_SERVER; // Do not run webpack dev server
-const SHOULD_FORK_TS = DEV_MODE && !env.NO_TS_FORK; // Do not run fork-ts plugin (or if not dev env)
+const SHOULD_FORK_TS = DEV_MODE && !env.NO_TS_FORK;
 const SHOULD_HOT_MODULE_RELOAD = DEV_MODE && !!env.SENTRY_UI_HOT_RELOAD;
-const SHOULD_LAZY_LOAD = DEV_MODE && !!env.SENTRY_UI_LAZY_LOAD;
+// const SHOULD_LAZY_LOAD = DEV_MODE && !!env.SENTRY_UI_LAZY_LOAD;
 
 // Deploy previews are built using vercel. We can check if we're in vercel's
 // build process by checking the existence of the PULL_REQUEST env var.
@@ -105,25 +93,6 @@ const SENTRY_SPA_DSN = SENTRY_EXPERIMENTAL_SPA ? env.SENTRY_SPA_DSN : undefined;
 const sentryDjangoAppPath = path.join(__dirname, 'src/sentry/static/sentry');
 const distPath = env.SENTRY_STATIC_DIST_PATH || path.join(sentryDjangoAppPath, 'dist');
 const staticPrefix = path.join(__dirname, 'static');
-
-// Locale file extraction build step
-if (env.SENTRY_EXTRACT_TRANSLATIONS === '1') {
-  babelConfig.plugins?.push([
-    'module:babel-gettext-extractor',
-    {
-      fileName: 'build/javascript.po',
-      baseDirectory: path.join(__dirname),
-      functionNames: {
-        gettext: ['msgid'],
-        ngettext: ['msgid', 'msgid_plural', 'count'],
-        gettextComponentTemplate: ['msgid'],
-        t: ['msgid'],
-        tn: ['msgid', 'msgid_plural', 'count'],
-        tct: ['msgid'],
-      },
-    },
-  ]);
-}
 
 // Locale compilation and optimizations.
 //
@@ -164,67 +133,39 @@ const localeToLanguage = (locale: string) => locale.toLowerCase().replace('_', '
 const supportedLocales = localeCatalog.supported_locales;
 const supportedLanguages = supportedLocales.map(localeToLanguage);
 
-type CacheGroups = Exclude<
-  NonNullable<Configuration['optimization']>['splitChunks'],
-  false | undefined
->['cacheGroups'];
-
-type CacheGroupTest = (
-  module: webpack.Module,
-  context: Parameters<webpack.optimize.SplitChunksPlugin['options']['getCacheGroups']>[1]
-) => boolean;
-
 // A mapping of chunk groups used for locale code splitting
-const localeChunkGroups: CacheGroups = {};
+const localeChunkGroups: Record<string, rspack.OptimizationSplitChunksCacheGroup> = {};
 
-supportedLocales
+for (const locale of supportedLocales) {
   // No need to split the english locale out as it will be completely empty and
   // is not included in the django layout.html.
-  .filter(l => l !== 'en')
-  .forEach(locale => {
-    const language = localeToLanguage(locale);
-    const group = `locale/${language}`;
+  if (locale === 'en') {
+    continue;
+  }
 
-    // List of module path tests to group into locale chunks
-    const localeGroupTests = [
-      new RegExp(`locale\\/${locale}\\/.*\\.po$`),
-      new RegExp(`moment\\/locale\\/${language}\\.js$`),
-    ];
+  const language = localeToLanguage(locale);
+  const group = `locale/${language}`;
 
-    // module test taken from [0] and modified to support testing against
-    // multiple expressions.
-    //
-    // [0] https://github.com/webpack/webpack/blob/7a6a71f1e9349f86833de12a673805621f0fc6f6/lib/optimize/SplitChunksPlugin.js#L309-L320
-    const groupTest: CacheGroupTest = (module, {chunkGraph}) =>
-      localeGroupTests.some(pattern =>
-        pattern.test(module?.nameForCondition?.() ?? '')
-          ? true
-          : chunkGraph.getModuleChunks(module).some(c => c.name && pattern.test(c.name))
-      );
-
-    // We are defining a chunk that combines the django language files with
-    // moment's locales as if you want one, you will want the other.
-    //
-    // In the application code you will still need to import via their module
-    // paths and not the chunk name
-    localeChunkGroups[group] = {
-      chunks: 'async',
-      name: group,
-      test: groupTest,
-      enforce: true,
-    };
-  });
-
-const babelOptions = {...babelConfig, cacheDirectory: true};
-const babelLoaderConfig = {
-  loader: 'babel-loader',
-  options: babelOptions,
-};
+  // We are defining a chunk that combines the django language files with
+  // moment's locales as if you want one, you will want the other.
+  //
+  // In the application code you will still need to import via their module
+  // paths and not the chunk name
+  localeChunkGroups[group] = {
+    chunks: 'async',
+    name: group,
+    test: new RegExp(
+      `(locale\\/${locale}\\/.*\\.po$)|(moment\\/locale\\/${language}\\.js$)`
+    ),
+    enforce: true,
+  };
+}
 
 /**
  * Main Webpack config for Sentry React SPA.
  */
-const appConfig: Configuration = {
+
+const appConfig: rspack.Configuration = {
   mode: WEBPACK_MODE,
   entry: {
     /**
@@ -247,6 +188,12 @@ const appConfig: Configuration = {
     sentry: 'less/sentry.less',
   },
   context: staticPrefix,
+  experiments: {
+    rspackFuture: {
+      disableTransformByDefault: true,
+    },
+    css: true,
+  },
   module: {
     /**
      * XXX: Modifying the order/contents of these rules may break `getsentry`
@@ -254,10 +201,32 @@ const appConfig: Configuration = {
      */
     rules: [
       {
-        test: /\.[tj]sx?$/,
-        include: [staticPrefix],
-        exclude: /(vendor|node_modules|dist)/,
-        use: babelLoaderConfig,
+        test: /\.(js|jsx|ts|tsx)$/,
+        exclude: /\/node_modules\//,
+        loader: 'builtin:swc-loader',
+        options: {
+          rspackExperiments: {
+            emotion: {
+              sourceMap: true,
+            },
+          },
+          sourceMap: true,
+          jsc: {
+            parser: {
+              syntax: 'typescript',
+              tsx: true,
+            },
+            target: 'es5', // this is a workaround because swc-emotion has a bug dealing with es5 tag template,see https://github.com/vercel/next.js/issues/38301
+            transform: {
+              react: {
+                runtime: 'automatic',
+                development: DEV_MODE,
+                refresh: DEV_MODE,
+                importSource: '@emotion/react',
+              },
+            },
+          },
+        },
       },
       {
         test: /\.po$/,
@@ -271,36 +240,22 @@ const appConfig: Configuration = {
       },
       {
         test: /\.pegjs/,
-        use: {loader: 'pegjs-loader'},
+        loader: 'pegjs-loader',
       },
       {
         test: /\.css/,
-        use: ['style-loader', 'css-loader'],
+        type: 'css',
       },
       {
         test: /\.less$/,
         include: [staticPrefix],
-        use: [
-          {
-            loader: MiniCssExtractPlugin.loader,
-            options: {
-              publicPath: 'auto',
-            },
-          },
-          'css-loader',
-          'less-loader',
-        ],
+        use: ['less-loader'],
+        type: 'css',
       },
       {
         test: /\.(woff|woff2|ttf|eot|svg|png|gif|ico|jpg|mp4)($|\?)/,
         type: 'asset',
       },
-    ],
-    noParse: [
-      // don't parse known, pre-built javascript files (improves webpack perf)
-      /jed\/jed\.js/,
-      /marked\/lib\/marked\.js/,
-      /terser\/dist\/bundle\.min\.js/,
     ],
   },
   plugins: [
@@ -310,63 +265,25 @@ const appConfig: Configuration = {
      */
     new SentryInstrumentation(),
 
-    // Do not bundle moment's locale files as we will lazy load them using
-    // dynamic imports in the application code
-    new webpack.IgnorePlugin({
-      contextRegExp: /moment$/,
-      resourceRegExp: /^\.\/locale$/,
-    }),
-
     /**
      * TODO(epurkhiser): Figure out if we still need these
      */
-    new webpack.ProvidePlugin({
+    new rspack.ProvidePlugin({
       process: 'process/browser',
       Buffer: ['buffer', 'Buffer'],
     }),
 
     /**
-     * Extract CSS into separate files.
-     */
-    new MiniCssExtractPlugin({
-      // We want the sentry css file to be unversioned for frontend-only deploys
-      // We will cache using `Cache-Control` headers
-      filename: 'entrypoints/[name].css',
-    }),
-
-    /**
      * Defines environment specific flags.
      */
-    new webpack.DefinePlugin({
-      'process.env': {
-        NODE_ENV: JSON.stringify(env.NODE_ENV),
-        IS_ACCEPTANCE_TEST: JSON.stringify(IS_ACCEPTANCE_TEST),
-        DEPLOY_PREVIEW_CONFIG: JSON.stringify(DEPLOY_PREVIEW_CONFIG),
-        EXPERIMENTAL_SPA: JSON.stringify(SENTRY_EXPERIMENTAL_SPA),
-        SPA_DSN: JSON.stringify(SENTRY_SPA_DSN),
-        SENTRY_RELEASE_VERSION: JSON.stringify(SENTRY_RELEASE_VERSION),
-      },
+    new rspack.DefinePlugin({
+      'process.env.NODE_ENV': JSON.stringify(env.NODE_ENV),
+      'process.env.IS_ACCEPTANCE_TEST': JSON.stringify(IS_ACCEPTANCE_TEST),
+      'process.env.DEPLOY_PREVIEW_CONFIG': JSON.stringify(DEPLOY_PREVIEW_CONFIG),
+      'process.env.EXPERIMENTAL_SPA': JSON.stringify(SENTRY_EXPERIMENTAL_SPA),
+      'process.env.SPA_DSN': JSON.stringify(SENTRY_SPA_DSN),
+      'process.env.SENTRY_RELEASE_VERSION': JSON.stringify(SENTRY_RELEASE_VERSION),
     }),
-
-    /**
-     * This removes empty js files for style only entries (e.g. sentry.less)
-     */
-    new FixStyleOnlyEntriesPlugin({verbose: false}),
-
-    ...(SHOULD_FORK_TS
-      ? [
-          new ForkTsCheckerWebpackPlugin({
-            typescript: {
-              configFile: path.resolve(__dirname, './config/tsconfig.build.json'),
-              configOverwrite: {
-                compilerOptions: {incremental: true},
-              },
-            },
-            devServer: false,
-            // memorylimit is configured in package.json
-          }),
-        ]
-      : []),
 
     /**
      * Restrict translation files that are pulled in through app/translations.jsx
@@ -376,13 +293,13 @@ const appConfig: Configuration = {
      * Without this, webpack will still output all of the unused locale files despite
      * the application never loading any of them.
      */
-    new webpack.ContextReplacementPlugin(
+    new ContextReplacementPlugin(
       /sentry-locale$/,
       path.join(__dirname, 'src', 'sentry', 'locale', path.sep),
       true,
       new RegExp(`(${supportedLocales.join('|')})/.*\\.po$`)
     ),
-    new webpack.ContextReplacementPlugin(
+    new ContextReplacementPlugin(
       /moment\/locale/,
       new RegExp(`(${supportedLanguages.join('|')})\\.js$`)
     ),
@@ -391,7 +308,7 @@ const appConfig: Configuration = {
      * Copies file logo-sentry.svg to the dist/entrypoints directory so that it can be accessed by
      * the backend
      */
-    new CopyPlugin({
+    new rspack.CopyRspackPlugin({
       patterns: [
         {
           from: path.join(staticPrefix, 'images/logo-sentry.svg'),
@@ -410,11 +327,6 @@ const appConfig: Configuration = {
             ]
           : []),
       ],
-    }),
-
-    WebpackReactSourcemapsPlugin({
-      mode: IS_PRODUCTION ? 'strict' : undefined,
-      debug: false,
     }),
   ],
 
@@ -449,6 +361,7 @@ const appConfig: Configuration = {
 
     modules: ['node_modules'],
     extensions: ['.jsx', '.js', '.json', '.ts', '.tsx', '.less'],
+    tsConfigPath: path.resolve(__dirname, './tsconfig.json'),
   },
   output: {
     crossOriginLoading: 'anonymous',
@@ -470,14 +383,14 @@ const appConfig: Configuration = {
       chunks: 'async',
       maxInitialRequests: 10, // (default: 30)
       maxAsyncRequests: 10, // (default: 30)
-      cacheGroups: {
-        ...localeChunkGroups,
-      },
+      cacheGroups: localeChunkGroups,
     },
 
     // This only runs in production mode
-    // Grabbed this example from https://github.com/webpack-contrib/css-minimizer-webpack-plugin
-    minimizer: ['...', new CssMinimizerPlugin()],
+    minimizer: [
+      new rspack.SwcCssMinimizerRspackPlugin(),
+      new rspack.SwcJsMinimizerRspackPlugin(),
+    ],
   },
   devtool: IS_PRODUCTION ? 'source-map' : 'eval-cheap-module-source-map',
 };
@@ -495,11 +408,11 @@ if (IS_TEST || IS_ACCEPTANCE_TEST) {
     'fixtures/integration-docs/_platforms.json'
   );
 } else {
-  const plugin = new IntegrationDocsFetchPlugin({basePath: __dirname});
-  appConfig.plugins?.push(plugin);
-  appConfig.resolve!.alias!['integration-docs-platforms'] = plugin.modulePath;
+  // const plugin = new IntegrationDocsFetchPlugin({basePath: __dirname});
+  // appConfig.plugins?.push(plugin);
+  // appConfig.resolve!.alias!['integration-docs-platforms'] = plugin.modulePath;
 }
-
+//
 if (IS_ACCEPTANCE_TEST) {
   appConfig.plugins?.push(new LastBuiltPlugin({basePath: __dirname}));
 }
@@ -514,24 +427,23 @@ if (
     // Hot reload react components on save
     // We include the library here as to not break docker/google cloud builds
     // since we do not install devDeps there.
-    const ReactRefreshWebpackPlugin = require('@pmmmwh/react-refresh-webpack-plugin');
-    appConfig.plugins?.push(new ReactRefreshWebpackPlugin());
+    appConfig.plugins?.push(new ReactRefreshRspackPlugin());
 
     // TODO: figure out why defining output breaks hot reloading
     if (IS_UI_DEV_ONLY) {
       appConfig.output = {};
     }
 
-    if (SHOULD_LAZY_LOAD) {
-      appConfig.experiments = {
-        lazyCompilation: {
-          // enable lazy compilation for dynamic imports
-          imports: true,
-          // disable lazy compilation for entries
-          entries: false,
-        },
-      };
-    }
+    // if (SHOULD_LAZY_LOAD) {
+    //   appConfig.experiments = {
+    //     lazyCompilation: {
+    //       // enable lazy compilation for dynamic imports
+    //       imports: true,
+    //       // disable lazy compilation for entries
+    //       entries: false,
+    //     },
+    //   };
+    // }
   }
 
   appConfig.devServer = {
@@ -551,8 +463,7 @@ if (
       watch: true,
     },
     host: SENTRY_WEBPACK_PROXY_HOST,
-    // Don't reload on errors
-    hot: 'only',
+    hot: SHOULD_HOT_MODULE_RELOAD,
     port: Number(SENTRY_WEBPACK_PROXY_PORT),
     devMiddleware: {
       stats: 'errors-only',
@@ -598,7 +509,7 @@ if (
     appConfig.devServer = {
       ...appConfig.devServer,
       static: {
-        ...(appConfig.devServer.static as object),
+        ...(appConfig.devServer!.static as object),
         publicPath: '/_static/dist/sentry',
       },
       // syntax for matching is using https://www.npmjs.com/package/micromatch
@@ -619,7 +530,15 @@ if (DEV_MODE) {
   appConfig.plugins?.push(
     new WebpackHookPlugin({
       onBuildStart: ['yarn run spotlight-sidecar'],
-    })
+    }) as unknown as rspack.RspackPluginInstance
+  );
+}
+
+if (SHOULD_FORK_TS) {
+  appConfig.plugins?.push(
+    new WebpackHookPlugin({
+      onBuildStart: ['yarn tsc -p ./config/tsconfig.build.json --watch --incremental'],
+    }) as unknown as rspack.RspackPluginInstance
   );
 }
 
@@ -739,7 +658,6 @@ if (IS_UI_DEV_ONLY || SENTRY_EXPERIMENTAL_SPA) {
    * This is currently used for PR deploy previews, where only the frontend
    * is deployed.
    */
-  const HtmlWebpackPlugin = require('html-webpack-plugin');
   appConfig.plugins?.push(
     new HtmlWebpackPlugin({
       // Local dev vs vercel slightly differs...
@@ -754,41 +672,26 @@ if (IS_UI_DEV_ONLY || SENTRY_EXPERIMENTAL_SPA) {
       window: {
         __SENTRY_DEV_UI: true,
       },
-    })
+    }) as unknown as rspack.RspackPluginInstance
   );
 }
 
-const minificationPlugins = [
+const minificationPlugins: rspack.RspackPluginInstance[] = [
   // This compression-webpack-plugin generates pre-compressed files
   // ending in .gz, to be picked up and served by our internal static media
   // server as well as nginx when paired with the gzip_static module.
-  //
-  // TODO(ts): The current @types/compression-webpack-plugin is still targeting
-  //           webpack@4, for now we just as any it.
   new CompressionPlugin({
     algorithm: 'gzip',
     test: /\.(js|map|css|svg|html|txt|ico|eot|ttf)$/,
-  }) as any,
-  // NOTE: In production mode webpack will automatically minify javascript
-  // using the TerserWebpackPlugin.
+  }) as unknown as rspack.RspackPluginInstance,
 ];
 
 if (IS_PRODUCTION) {
-  // NOTE: can't do plugins.push(Array) because webpack/webpack#2217
-  minificationPlugins.forEach(plugin => appConfig.plugins?.push(plugin));
+  appConfig.plugins?.push(...minificationPlugins);
 }
 
 // Cache webpack builds
 if (env.WEBPACK_CACHE_PATH) {
-  appConfig.cache = {
-    type: 'filesystem',
-    cacheLocation: path.resolve(__dirname, env.WEBPACK_CACHE_PATH),
-    buildDependencies: {
-      // This makes all dependencies of this file - build dependencies
-      config: [__filename],
-      // By default webpack and loaders are build dependencies
-    },
-  };
+  appConfig.cache = true;
 }
-
 export default appConfig;