--- a/static/less/auth.less
+++ b/static/less/auth.less
@@ -42,55 +42,55 @@ section.org-login {
   background-size: cover;
 
   &.google {
-    background-image: url(~sentry-logos/logo-google.svg);
+    background-image: url(sentry-logos/logo-google.svg);
   }
 
   &.github {
-    background-image: url(~sentry-logos/logo-github.svg);
+    background-image: url(sentry-logos/logo-github.svg);
   }
 
   &.onelogin {
-    background-image: url(~sentry-logos/logo-onelogin.svg);
+    background-image: url(sentry-logos/logo-onelogin.svg);
   }
 
   &.auth0 {
-    background-image: url(~sentry-logos/logo-auth0.svg);
+    background-image: url(sentry-logos/logo-auth0.svg);
   }
 
   &.okta {
-    background-image: url(~sentry-logos/logo-okta.svg);
+    background-image: url(sentry-logos/logo-okta.svg);
   }
 
   &.jumpcloud {
-    background-image: url(~sentry-logos/logo-jumpcloud.svg);
+    background-image: url(sentry-logos/logo-jumpcloud.svg);
   }
 
   &.rippling {
-    background-image: url(~sentry-logos/logo-rippling.svg);
+    background-image: url(sentry-logos/logo-rippling.svg);
   }
 
   &.active-directory {
-    background-image: url(~sentry-logos/logo-microsoft.svg);
+    background-image: url(sentry-logos/logo-microsoft.svg);
   }
 
   &.saml2 {
-    background-image: url(~sentry-logos/logo-saml2.svg);
+    background-image: url(sentry-logos/logo-saml2.svg);
   }
 
   &.slack {
-    background-image: url(~sentry-logos/logo-slack.svg);
+    background-image: url(sentry-logos/logo-slack.svg);
   }
 
   &.vsts {
-    background-image: url(~sentry-logos/logo-azure.svg);
+    background-image: url(sentry-logos/logo-azure.svg);
   }
 
   &.fly {
-    background-image: url(~sentry-logos/logo-fly.svg);
+    background-image: url(sentry-logos/logo-fly.svg);
   }
 
   &.discord {
-    background-image: url(~sentry-logos/logo-discord.svg);
+    background-image: url(sentry-logos/logo-discord.svg);
   }
 }