--- a/static/less/layout.less
+++ b/static/less/layout.less
@@ -320,7 +320,7 @@ body.auth {
 */
 
 .pattern-bg {
-  background: #f2f1f3 url(../images/pattern/sentry-pattern.png);
+  background: #f2f1f3 url(sentry-images/pattern/sentry-pattern.png);
   background-size: 340px;
   position: fixed;
   top: 0;