## Summary

This pull-request removes babel and Webpack in favor of Rspack and SWC. This change comes with a significant performance boost and improves our productivity. Currently, it takes 14 seconds to build Sentry. Prior to this PR, it took 58 seconds.

## Before

We are using Babel and Webpack for building our frontend application. It took almost 58 seconds on my M2 machine to run `yarn build-production` command.

```
webpack 5.87.0 compiled with 1 warning in 56351 ms
✨  Done in 57.91s.
```

## After

We removed 14 dependencies and added 3 new dependencies. Depending on Babel and Jest, we can remove more. Currently it takes 14 seconds to build, minify both JS and CSS files.

```
Rspack compiled successfully in 12.90 s
✨  Done in 14.00s.
```

## TODOs:

- [ ] In order to not break the Vercel previews we need the following change of Rspack.
  - https://github.com/web-infra-dev/rspack/issues/4977
- [ ] Rspack doesn't support lazy compilation, but [it is in their roadmap.](https://www.rspack.dev/misc/roadmap.html#lazy-compilation-support)
- [ ] Research why we need `build-js-loader.ts` script since it brings an unnecessary dependency: `terser`
- [ ] Research the relation of Babel and Jest and remove unnecessary configs from `babel.config`.
- [ ] Make `tsc` child process DX a little bit better.
- [ ] Fix error from `downsample` npm library in production build
  - https://github.com/web-infra-dev/rspack/issues/4980 

Thanks @hardfist for helping with this PR!