@hardfist 2023-12-09
@anonrig do you know how to setup devenv for outside contributors, I'm not sure how should i setup the environment, I can't launch the server so I can't verify whether the migration runs successfully
![image](https://github.com/getsentry/sentry/assets/8898718/ce100523-39c1-46bb-9ff8-e37ce75d09bc)