@anonrig 2023-12-09
@hardist You don't need devenv for this repository. Cloning, installing dependencies and running "yarn dev-ui" is sufficient for running the frontend.