@hardfist 2023-12-09
> @Hardist You don't need devenv for the frontend app. Cloning, installing dependencies and running "yarn dev-ui" is sufficient for running the frontend.
> 
> The public documentation is available at https://develop.sentry.dev/environment/

no need for homepage but seems needed for login page