@anonrig 2023-12-09
> no need for homepage but seems needed for login page，but i will verify homepage first

You can use mkert command in package.json with cookie sync chrome extension. https://chrome.google.com/webstore/detail/sentry-cookie-sync/kchlmkcdfohlmobgojmipoppgpedhijh