@scttcper 2024-09-06
@ScriptedAlchemy @hardfist thanks for your time looking at this. I've opened #77077 to startup the investigation with rspack v1. Congrats on the launch btw